refactor: moved the consistent hash methods to its correct location (I wonder why did I put the methods in the cluster manager file)
