+++ internal/cluster/hashing.go
@@ -1,6 +1,10 @@
 package cluster
 
 import (
+	"hash/crc32"
+	"slices"
+	"sort"
+	"strconv"
 	"sync"
 )
 
@@ -16,4 +20,40 @@ func NewConsistentHash(replicas int) *ConsistentHash {
 		replicas: replicas,
 		hashMap:  make(map[uint32]string),
 	}
+}
+
+func (c *ConsistentHash) AddNode(node string) {
+	c.mu.Lock()
+	defer c.mu.Unlock()
+
+	for idx := range c.replicas {
+		strIdx := strconv.Itoa(idx)
+		vKey := node + "#" + strIdx
+
+		hash := crc32.ChecksumIEEE([]byte(vKey))
+		c.keys = append(c.keys, hash)
+		c.hashMap[hash] = node
+	}
+
+	slices.Sort(c.keys)
+}
+
+func (c *ConsistentHash) GetNode(key string) string {
+	c.mu.RLock()
+	defer c.mu.RUnlock()
+
+	if len(c.keys) == 0 {
+		return ""
+	}
+
+	hash := crc32.ChecksumIEEE([]byte(key))
+	nodeIdx := sort.Search(len(c.keys), func(i int) bool {
+		return c.keys[i] >= hash
+	})
+
+	if nodeIdx == len(c.keys) {
+		nodeIdx = 0
+	}
+
+	return c.hashMap[c.keys[nodeIdx]]
 }
